+++ lib/src/rsa.dart
@@ -6,28 +6,27 @@
 
 part of 'steel_crypt_base.dart';
 
-///RSA asymmetric encryption machine
+///RSA asymmetric encryption machine.
 class RsaCrypt {
-
-  ///Pair of private keys
+  ///Pair of private keys.
   var pair;
 
-  ///Construct with keys
+  ///Construct with keys.
   RsaCrypt ()  {
     pair = getRsaKeyPair(getSecureRandom());
   }
 
-  ///access private key
-  PrivateKey get privKey {
+  ///Access private key.
+  PrivateKey get randPrivKey {
     return pair.privateKey;
   }
 
-  ///access public key
-  PublicKey get pubKey {
+  ///Access public key.
+  PublicKey get randPubKey {
     return pair.publicKey;
   }
 
-  ///create Random param for RSA keypair
+  ///create Random param for RSA keypair.
   static SecureRandom getSecureRandom() {
     var secureRandom = FortunaRandom();
     var random = Random.secure();
@@ -39,7 +38,7 @@ class RsaCrypt {
     return secureRandom;
   }
 
-  ///create RSA keypair given SecureRandom
+  ///Create RSA keypair given SecureRandom.
   static AsymmetricKeyPair<PublicKey, PrivateKey> getRsaKeyPair(
       SecureRandom secureRandom) {
     var rsapars = RSAKeyGeneratorParameters(BigInt.from(65537), 2048, 5);
@@ -49,7 +48,15 @@ class RsaCrypt {
     return keyGenerator.generateKeyPair();
   }
 
-  ///encrypt using RSA
+  ///Parse key from PEM file
+  Future<T> parseKeyFromFile<T extends RSAAsymmetricKey>(String filename) async {
+    final file = File(filename);
+    final key = await file.readAsString();
+    final parser = RSAKeyParser();
+    return parser.parse(key) as T;
+  }
+
+  ///Encrypt using RSA.
   String encrypt(String text, RSAPublicKey pubKey) {
     var cipher = OAEPEncoding(RSAEngine());
     cipher.init(true, PublicKeyParameter<RSAPublicKey>(pubKey));
@@ -57,11 +64,82 @@ class RsaCrypt {
     return base64Encode(output1);
   }
 
-  ///decrypt using RSA
-  String decrypt(String input, RSAPrivateKey privateKey) {
+  ///Decrypt using RSA.
+  String decrypt(String encrypted, RSAPrivateKey privateKey) {
     var cipher = OAEPEncoding(RSAEngine());
     cipher.init(false, PrivateKeyParameter<RSAPrivateKey>(privateKey));
-    Uint8List output = cipher.process(base64Decode(input));
+    Uint8List output = cipher.process(base64Decode(encrypted));
     return utf8.decode(output);
+  }
+}
+
+/// RSA PEM parser. Not for public use, don't use!!!
+class RSAKeyParser {
+  RSAAsymmetricKey parse(String key) {
+    final rows = key.split(RegExp(r'\r\n?|\n'));
+    final header = rows.first;
+
+    if (header == '-----BEGIN RSA PUBLIC KEY-----') {
+      return _parsePublic(_parseSequence(rows));
+    }
+
+    if (header == '-----BEGIN PUBLIC KEY-----') {
+      return _parsePublic(_pkcs8PublicSequence(_parseSequence(rows)));
+    }
+
+    if (header == '-----BEGIN RSA PRIVATE KEY-----') {
+      return _parsePrivate(_parseSequence(rows));
+    }
+
+    if (header == '-----BEGIN PRIVATE KEY-----') {
+      return _parsePrivate(_pkcs8PrivateSequence(_parseSequence(rows)));
+    }
+
+    throw FormatException('Unable to parse key, invalid format.', header);
+  }
+
+  RSAAsymmetricKey _parsePublic(ASN1Sequence sequence) {
+    final modulus = (sequence.elements[0] as ASN1Integer).valueAsBigInteger;
+    final exponent = (sequence.elements[1] as ASN1Integer).valueAsBigInteger;
+
+    return RSAPublicKey(modulus, exponent);
+  }
+
+  RSAAsymmetricKey _parsePrivate(ASN1Sequence sequence) {
+    final modulus = (sequence.elements[1] as ASN1Integer).valueAsBigInteger;
+    final exponent = (sequence.elements[3] as ASN1Integer).valueAsBigInteger;
+    final p = (sequence.elements[4] as ASN1Integer).valueAsBigInteger;
+    final q = (sequence.elements[5] as ASN1Integer).valueAsBigInteger;
+
+    return RSAPrivateKey(modulus, exponent, p, q);
+  }
+
+  ASN1Sequence _parseSequence(List<String> rows) {
+    final keyText = rows
+        .skipWhile((row) => row.startsWith('-----BEGIN'))
+        .takeWhile((row) => !row.startsWith('-----END'))
+        .map((row) => row.trim())
+        .join('');
+
+    final keyBytes = Uint8List.fromList(base64.decode(keyText));
+    final asn1Parser = ASN1Parser(keyBytes);
+
+    return asn1Parser.nextObject() as ASN1Sequence;
+  }
+
+  ASN1Sequence _pkcs8PublicSequence(ASN1Sequence sequence) {
+    final ASN1Object bitString = sequence.elements[1];
+    final bytes = bitString.valueBytes().sublist(1);
+    final parser = ASN1Parser(Uint8List.fromList(bytes));
+
+    return parser.nextObject() as ASN1Sequence;
+  }
+
+  ASN1Sequence _pkcs8PrivateSequence(ASN1Sequence sequence) {
+    final ASN1Object bitString = sequence.elements[2];
+    final bytes = bitString.valueBytes();
+    final parser = ASN1Parser(bytes);
+
+    return parser.nextObject() as ASN1Sequence;
   }
 }